Changed movement system to states instead of booleans, added attack hitboxes
Changed booleans in PlayerMovement to an Enum "States", added an attack prefab with hitboxes and an attackDamage script controlling it,
Added a sandbag prefab to have for the prototype. Added health script to keep track of player health. Improved some animations. Fixed animation transitions
to use playerState from playerMovement instead of a million different variables.